Avoiding Wordpress notice loading enqueueables
If we try to load enqueuables before certain hooks, Wordpress gets crazy
and notices the following:

Notice: wp_enqueue_script was called <strong>incorrectly</strong>.
Scripts and styles should not be registered or enqueued until the
<code>wp_enqueue_scripts</code>, <code>admin_enqueue_scripts</code>, or
<code>login_enqueue_scripts</code> hooks. This notice was triggered by
the <code>script-main</code> handle. Please see
<a href="https://wordpress.org/support/article/debugging-in-wordpress/">
Debugging in WordPress</a> for more information. (This message was added
in version 3.3.0.) in /var/www/wp/wp-core/wp-includes/functions.php on
line 5775