Refactoring enqueueables classes
- relativeFilepath method discontinued
- filename and mountFileUrl methods introduced

With the changes above we may now define only the asset filename (script
or style) and it's already known how it should mount the URL to this
file. However, as we needed to framework development, you may also
overwrite the mountFileUrl method to change the entire URL to your file.