Trying to workaround a WP-CLI bug
The following commands from WP CLI does not work with the --format flag
passing any valid value (including csv):

- profile stage bootstrap
- profile stage main_query
- profile stage template

Tested with csv and json format values. The following error occurs:

Error: Invalid field: callback_count.

To workaround it we now use the above commands with no format which
leads to a table mounted as TSV. So we just introduced a new
writeTableFromTsv method.