Fixing frontend error
The previous code always lead us to the following error: Uncaught
ReferenceError: wp is not defined

That occurs mainly because of an inline script associated after the WP
default script of id 'wp-i18n-js'. But if only ignore scripts with
associated after inline scripts from being deferred, more errors are
raised because of scripts wrong concurrency loading.

That's why we must also let the scripts from wp-includes (default from
WP) be controlled by it. We can't know if they are prepared to be
defered or not.