Fixing stubborn WordPress options
Raw queries must be used instead of
$this->runWpCliCommandSilentlyWithoutInterruption("option update siteurl
$siteurl_option_value"); and
$this->runWpCliCommandSilentlyWithoutInterruption("option update home
$home_option_value"); because WP-CLI let those options be retrieved from
WP_SITEURL and WP_HOME constants when checking if it should change them
in database leading to a wrong results of "Success: Value passed for
'siteurl' option is unchanged." and "Success: Value passed for 'home'
option is unchanged."

Also, for some reason, even with the constants defined, WordPress does
not work correctly if the database isn't correct either. (even saying
those options won't be retrieved from database) We could test it when
running multiple environments and needed to access a URL with a defined
custom port like https://another-project.test:4430, which would redirect
us to the other project url like https://project.test even with all the
right configurations.