Fixing database connection error on wp-cli commands
The main problem was to keep APP_URL as an empty string, which leads to
empty siteurl and homeurl in wp_options table. this is onvalid to
WordPress, but the error raised just says nothing more than the
following:

Error: One or more database tables are unavailable. The database may
need to be repaired.

Also we removed env.php and set those variables as defaults in the
customized framework wp-config.php